requirements.txt
accelerate==1.10.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.46
beautifulsoup4==4.13.5
bitsandbytes==0.46.1
bleach==6.2.0
Brotli==1.1.0
build==1.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
conda-pack==0.8.1
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.6
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.16
decorator==4.4.2
decord==0.6.0
deepspeed @ file:///data1/pengrui/Download/DeepSpeed-0.17.5/dist/deepspeed-0.17.5%2Bunknown-py3-none-any.whl#sha256=821e566629306dcde4c3710ab38a9cca9d8e526fc3186bb7da3a57353fd2c526
defusedxml==0.7.1
dill==0.3.8
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
fastjsonschema==2.21.2
ffmpy==0.6.1
filelock==3.14.0
flash_attn @ file:///data1/pengrui/Download/flash_attn-2.8.3%2Bcu12torch2.8cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=7aa420a0f6a7fb0a75d3d4f62f7e3a41b51dba7f1fee0defaed051c5465ccb32
fonttools==4.59.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
future==1.0.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.8
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
-e git+https://github.com/open-mmlab/mmcv.git@90d83c94cfb967ef162c449faf559616f31f28c2#egg=mmcv
mmengine==0.10.6
model-index==0.1.11
modelscope==1.29.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.1.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
omegaconf==2.3.0
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.11.3
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
portalocker==3.2.0
prettytable==3.16.0
proglog==0.1.12
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.4.2
rpds-py==0.27.1
ruff==0.12.11
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.47.3
sty==1.0.6
sympy==1.13.3
tabulate==0.9.0
termcolor==3.1.0
terminado==0.18.1
tifffile==2025.5.10
tiktoken==0.11.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.22.0
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tornado==6.5.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.56.1
transformers-stream-generator==0.0.5
triton==3.4.0
typer==0.17.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.35.0
validators==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
xlsxwriter==3.2.9
xtuner==0.2.0
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
